fix: increase minimum_hours to 48
With minimum_hours=24, the bot ends up comparing forecasts from H+24 to H+0, but forecasts made at H+0 tend to change a lot and not be very meaningful. This leads to "false alerts" such as https://twitter.com/MetaculusAlert/status/1500713083878232069 or https://twitter.com/MetaculusAlert/status/1500713035786244099. This ensures that the first comparison is made at H+48, compared to H+24.